create folder for replay related code
