+++ dogfight/src/replay/file.rs
@@ -12,8 +12,8 @@ use crate::{
 #[derive(Serialize, Deserialize, Debug, TS)]
 #[ts(export)]
 pub struct ReplayTickCommand {
-    player_guid_index: u16,
-    command: PlayerCommand,
+    pub player_guid_index: u16,
+    pub command: PlayerCommand,
 }
 
 impl ReplayTickCommand {
@@ -28,8 +28,8 @@ impl ReplayTickCommand {
 #[derive(Serialize, Deserialize, Debug, TS)]
 #[ts(export)]
 pub struct ReplayTick {
-    tick_number: u32,
-    commands: Vec<ReplayTickCommand>,
+    pub tick_number: u32,
+    pub commands: Vec<ReplayTickCommand>,
 }
 
 impl ReplayTick {
@@ -73,95 +73,6 @@ impl ReplayFile {
             player_guids: BTreeMap::new(),
             ticks: vec![],
         }
-    }
-}
-
-impl NetworkedBytes for ReplayFile {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut bytes = vec![];
-
-        // look how pretty this is
-        bytes.extend(self.player_guids.to_bytes());
-        bytes.extend(self.ticks.to_bytes());
-        bytes.extend(self.level_name.to_bytes());
-        bytes.extend(self.level_data.to_bytes());
-        bytes.extend(self.build_version.to_bytes());
-        bytes.extend(self.build_commit.to_bytes());
-
-        bytes
-    }
-
-    fn from_bytes(bytes: &[u8]) -> Option<(&[u8], Self)>
-    where
-        Self: Sized,
-    {
-        let (bytes, player_guids) = BTreeMap::<String, u16>::from_bytes(bytes)?;
-        let (bytes, ticks) = Vec::<ReplayTick>::from_bytes(bytes)?;
-        let (bytes, level_name) = String::from_bytes(bytes)?;
-        let (bytes, level_data) = String::from_bytes(bytes)?;
-        let (bytes, build_version) = String::from_bytes(bytes)?;
-        let (bytes, build_commit) = String::from_bytes(bytes)?;
-
-        Some((
-            bytes,
-            Self {
-                build_version,
-                build_commit,
-                level_name,
-                level_data,
-                player_guids,
-                ticks,
-            },
-        ))
-    }
-}
-
-impl NetworkedBytes for ReplayTick {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut bytes = vec![];
-        bytes.extend(u32::to_bytes(&self.tick_number));
-        bytes.extend(Vec::<ReplayTickCommand>::to_bytes(&self.commands));
-        bytes
-    }
-
-    fn from_bytes(bytes: &[u8]) -> Option<(&[u8], Self)>
-    where
-        Self: Sized,
-    {
-        let (mut bytes, tick_number) = u32::from_bytes(bytes)?;
-        let (slice, commands) = Vec::<ReplayTickCommand>::from_bytes(bytes)?;
-        bytes = slice;
-        Some((
-            bytes,
-            Self {
-                tick_number,
-                commands,
-            },
-        ))
-    }
-}
-
-impl NetworkedBytes for ReplayTickCommand {
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut bytes = vec![];
-        bytes.extend(self.player_guid_index.to_bytes());
-        bytes.extend(self.command.to_bytes());
-        bytes
-    }
-
-    fn from_bytes(bytes: &[u8]) -> Option<(&[u8], Self)>
-    where
-        Self: Sized,
-    {
-        let (bytes, player_guid_index) = u16::from_bytes(bytes)?;
-        let (bytes, command) = PlayerCommand::from_bytes(bytes)?;
-        Some((
-            bytes,
-            Self {
-                player_guid_index,
-                command,
-            },
-        ))
     }
 }
